Don't display integrated-vtysh-config as its the default in CL.
Ticket: CM-11832
Reviewed By:
Testing Done: Testing that its not displayed if enabled & only if disabled

In the spirit of not displaying the defaults, we shouldn't display
"service integrated-vtysh-config" as its the default. It also tends to
clutter the output with stuff the user doesn't know or care about. This
patch removes displaying that and only prints it when the option is
disabled.