bgpd: Pass correct bgp-name for clear command
Found that original fix for CM-10113 had a significant flaw, that
by deriving the bgp instance from the vty->index, if a clear ip bgp
command was entered from a vty index other then bgp, a crash occurred.
This fix passes the bgp->name to the clear functions so the correct
instance can be derived.  Tested manually in both the CM-10113 case
as well as the problem discovered while testing CM-11480.

Ticket: CM-10113
Signed-off-by: Don Slice
Reviewed-by: