Remove individual daemon conf files and replace with Quagga.conf
Ticket: CM-11777
Reviewed By: CCR-5096
Testing Done:

The recommended, and in many ways the only supported, model for
the configuration file of quagga is to use a single Quagga.conf
configuration file. However, we weren't shipping with this model,
which led to some confusion amongst users. This patch fixes this
by removing all individual daemon configuration files and replacing
it with the single Quagga.conf file.