bgpd: Display interface next-hop for "show ip bgp" with unnumbered
Found that the logic had been changed to determine whether the next-hop
is a v4 or v6 address.  This caused an unnumbered interface to be seen
as ipv4 instead of ipv6 so the swp port was not correctly displayed.
Changed it back. Manual testing attaced to the ticket and bgp-min will
be run before committing.

Ticket: CM-12759
Signed-off-by: Don Slice
Reviewed-by: CCR-5166