lib: apply mask to prefix in prefix-list
A crash occurred if a prefix was defined in a prefix-list that
contained bits in the prefix but a /0 mask.  Resolving that crash
and improving usability by applying the mask to the supplied prefix
and notifying the user if the prefix was modified.

Ticket: CM-12744
Signed-off-by: Don Slice
Reviewed_By:
Testing Done: Manual testing attached to the ticket, bgp-min, bgp-smoke
ospf-min, and ospf-smoke all completed before commit