zserv: always send all information about each route
Most routing daemons are not interested in certain pieces of information
when a redistributed route is being removed, like its metric and distance.

ldpd, in the other hand, needs to know the distance of the removed routes
in order to work properly. Now, instead of adding another exception in
zserv's code for ldpd, let's make zebra always send all information
about each route to its clients, independently if the route is being
added or removed. This is ok because all daemons are already prepared
to process these additional fields when the appropriate flags are set
in the zebra messages.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>