zebra: Move netlink error message under a debug
In some circumstances, the quagga log is being filled with repetitive
error messages reporting "network is down" with RTM_NEWROUTE.  Moved this
particular scenario under "debug zebra kernel" instead of making it an
unprotected error message.  Manually tested using the same script with and
without the fix to verify the message is suppressed.

Ticket: CM-11173
Signed-off-by: Don Slice
Reviewed-by: Donald Sharp